[Sermed-Data] Test de integracion de afiliado listos

+++ Sermed-Data/src/test/java/ar/com/koodi/sermeddata/IntegrationTest/AfeccionRepositorioIntegrationTest.java
@@ -50,7 +50,7 @@ public class AfeccionRepositorioIntegrationTest {
 	@Test
 	public void findByNombreUnico_ExistenAfeccion_DevuelveAfeccion() {
 		Afeccion resultado = this.afeccionRepositorioImplementacion.findByNombreUnico("Gripe");
-		Assert.assertEquals(2, resultado.getIdAfeccion());
+		Assert.assertEquals(2, resultado.getIdAfeccion().intValue());
 	}
 
 	@Test
